// src/services/admin/saree-catalog/type.ts
export interface ViewSareeCatalogData {
  _id: string;
  productName: string;
  brandName: string;
  title: string;
  description: string;
  color: string;
  price: number;
  discountPrice: number;
  styleCode: string;
  pattern: string;
  packOf: number;
  occasion: string;
  decorativeMaterial: string;
  constructionType: string;
  fabricCare: string;
  otherDetails: string;
  fabric: string;
  type: string;
  blousePiece: string;
  sariStyle: string;
  netQuantity: number;
  sariLength: number;
  blousePieceLength: number;
  weight: number;
  imageUrl: string;
  createdAt: string;
}